% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{compute_nrsme}
\alias{compute_nrsme}
\title{A wrapper function of \code{nrmse_idf}, computes NRMSE same for a range of durations}
\usage{
compute_nrsme(
  station_data,
  declustering_duration,
  kappa_fit,
  sigma_fit,
  xi_fit,
  init_time_step = 1,
  q = 0
)
}
\arguments{
\item{station_data}{data_frame of observation for a station (nobs x aggregation duration)}

\item{declustering_duration}{a vector same length as \code{duration} .Whether the data should be temporally declustered, if yes the time step for each duration, or a vector of 1 otherwise}

\item{kappa_fit}{a numerical vector  of  egpd 'kappa'  parmater. Same length as \code{ncol(station_data)}}

\item{sigma_fit}{same as  \code{kappa_fit} but for egpd 'sigma'}

\item{xi_fit}{same as  \code{kappa_fit} but for egpd shape parammter (xi)}

\item{init_time_step}{a scalar, eg 1 or 2. The time step to start declustering the data. eg, for hourly data, if  \code{declustering =3} and  \code{init_time_step = 2}, then the 2nd hour will be selected, and then a sequence is applied}

\item{q}{optional. a number in percentage, or any character eg \code{nrsme_quantile = 90}, the normalized root mean square error will only be computed on excesses of the quantile. Quantile is over ALL  values (zeros included)
If a character, then a weighted normalized root mean square error will be used.}
}
\value{
a numerical vector of 'nrmse'
}
\description{
NRMSE stands for Normalized Root Mean Square Error
}
\details{
to be added
}
