% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpdIDF.R
\docType{package}
\name{egpdIDF}
\alias{egpdIDF}
\title{Intensity Duration Frequency model (IDF) using the Extended Generalised Pareto Distribution}
\description{
IDF curves are usually modelled using Generalized Exteme Value distributions (GEV).
Here, we consider the three parameter EGPD of Naveau et al 2016 to bulid IDF curves. The package provides the implemenattion
of 10 diffent IDF modelling approaches as contained in Haruna et al 2023.
}
\author{
Abubakar HARUNA \email{abbaharuna87@gmail.com}
}
