% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egpdIDF.R
\docType{data}
\name{precipdata}
\alias{precipdata}
\title{A fictious hourly  time series of precipitation}
\format{
## `precipdata`
A data frame of two columns and 1577952 rows:
\describe{
  \item{date}{date}
  \item{SCH}{precipitation amounts}
}
}
\usage{
precipdata
}
\description{
A fictious hourly  time series of precipitation
}
\keyword{datasets}
