% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{qqplot_egpd}
\alias{qqplot_egpd}
\title{Quantile-Quantile Plot of EGPD model}
\usage{
qqplot_egpd(
  data,
  kappa,
  sigma,
  xi,
  ranking_method = "average",
  title = "QQPLOT EGPD",
  subtitle = "",
  confid = FALSE,
  q_up = NA,
  q_low = NA
)
}
\arguments{
\item{data}{vector of observation for a station}

\item{kappa}{egpd 'kappa'  paramater}

\item{sigma}{egpd 'sigma' paramater}

\item{xi}{egpd shape parammter (xi)}

\item{ranking_method}{a character string specifying how ties are treated. see \code{rank}  function in \code{R}. Defualts to "average"}

\item{title}{optional a character string for the title of the plot}

\item{subtitle}{optional a character string for the subtitle of the plot}

\item{confid}{logical, whether confidence bounds are to be plotted, defaults to \code{FALSE}}

\item{q_up}{a numerical vector. For the upper confidence levels, only if \code{confid = T}}

\item{q_low}{a numerical vector. For the lower confidence levels, only if \code{confid = T}}
}
\value{
a ggplot object
}
\description{
The function plot qqplot from a fitted egpd model
}
\details{
Note that  for confidence bounds to be plotted, \code{q_up} and \code{q_low} have to be separately computed by the user, eg via bootstrap
}
