% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{trplot_egpd}
\alias{trplot_egpd}
\title{Quantile-Quantile Plot of EGPD model}
\usage{
trplot_egpd(data, kappa, sigma, xi, npy, max_T = 1000)
}
\arguments{
\item{data}{vector of observation for a station}

\item{kappa}{egpd 'kappa'  paramater}

\item{sigma}{egpd 'sigma' paramater}

\item{xi}{egpd shape parammter (xi)}

\item{npy}{number of observations per year. Eg if  \code{data} contains only daily data in summer, then \code{npy = 92}. If  for a year, then \code{npy = 365.25} etc description}

\item{max_T}{a scalar. The largest return level to be considered in the plot}
}
\value{
a ggplot object
}
\description{
The function plot qqplot from a fitted egpd model
}
\details{
to be done
}
