% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_functions.R
\name{egpd_idf_init}
\alias{egpd_idf_init}
\title{Function to estimates the initial parameters for the IDF_egpd models}
\usage{
egpd_idf_init(
  station_data,
  durations,
  declustering_duration,
  init_time_step = 1,
  fitting_method = "mle",
  auto_fit = T,
  nrmse_tol = 0.1,
  simple_scaling = T,
  use_r_optim = F,
  nrsme_quantile = 0
)
}
\arguments{
\item{station_data}{data_frame of observation for a station (nobs x aggregation duration)}

\item{durations}{a (vector) of durations (same length as ncol of station_data)}

\item{declustering_duration}{a vector same length as \code{duration} .Whether the data should be temporally declustered, if yes the time step for each duration, or a vector of 1 otherwise}

\item{init_time_step}{a scalar, eg 1 or 2. The time step to start declustering the data. eg, for hourly data, if  \code{declustering =3} and  \code{init_time_step = 2}, then the 2nd hour will be selected, and then a sequence is applied}

\item{fitting_method}{eiither \code{"mle"} for maximum likelihood or \code{"pwm"} for probability weighted momoments. Defaults to \code{"mle"}}

\item{auto_fit}{logical, whether an automatic fit is to be done to find a censoring threshold that minimizes the normalized root mean square error; defaults to \code{TRUE}}

\item{nrmse_tol}{a scalar, defaults to \code{0.1}. defaines a relatively good egpd fit. If higher, and \code{auto_fit=TRUE}, automatic fit is done}

\item{simple_scaling}{logical, defaults to \code{TRUE}, defines the initial values values that would be returned}

\item{use_r_optim}{logical, defaults  to \code{FALSE}, if yes, \code{optim} will be used to find the best lower censoring values, otherwise, descrete values will be tested}

\item{nrsme_quantile}{a number in percentage, or any character eg \code{nrsme_quantile = 90}, the normalized root mean square error will only be computed on excesses of the quaneile.  If a charanter, then a weighted normalized root mean square error will be used.}
}
\value{
# A list:
   init: the initial parameters to use for fitting a egpd IDF model
   fits: egdp parameters fit for each duration seperately
   regres_summary: summary of the lm fit of the sigma vs duration
}
\description{
each duration is fitted separately, then lm is fitted to estimate the parameters of the IDF (sigma0 as the intecept, etha as the -ve of the slope)
 kappa and xi as the median of the d_ estimates
}
\details{
to be added
}
